Resolve TODOs
Enhanced 'parse_integer()' function: It now returns 1 if parsing
was successful, and 0 otherwise.

Optimized HTTP major and minor version parsing: Since these numbers
are currently (and are expected to remain) single-digit, calling
'parse_integer()' for them was overkill. A more efficient solution
using simple ASCII character subtraction has been implemented.

Fixed POST method parsing.

+++ ahttp_parser.h
@@ -83,7 +83,7 @@ struct http_parser {
     short int http_major;
     short int http_minor;
 
-    short int status; // only response
+    int status; // only response
     enum http_method method; // only request
 
     unsigned int error: 1;
